% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd.R
\name{cd}
\alias{cd}
\title{Change working directory using the cd syntax}
\usage{
cd(x = NULL)
}
\arguments{
\item{x}{A string giving the directory relative to the current working directory.}
}
\value{
None.
}
\description{
The function enables changing working directory using the cd syntax.
}
\note{
None.
}
\examples{
\dontrun{

system('mkdir test')

cd('test')

}
}
\references{
None
}
\author{
Xia Shen
}
