% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colVars.R
\name{colVars}
\alias{colVars}
\title{The function calculates the variance of each column of your data table}
\usage{
colVars(x)
}
\arguments{
\item{X}{A data.frame or matrix you want to calculate the variance of each column}
}
\value{
A vecotor of variance of each column.
}
\description{
The function calculates the variance of each column of your data table
}
\examples{

df <- matrix(rnorm(10,0,1),5,2)
df
res <- colVars(df)

}
\seealso{
\code{colSums}
\code{colMeans}
}
\author{
Ting Li, Xia Shen
}
