% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileRnw.R
\name{compileRnw}
\alias{compileRnw}
\title{Compile an Rmarkdown source document}
\usage{
compileRnw(fn)
}
\arguments{
\item{fn}{Rnw file name with path.}
}
\value{
A PDF document will be created in the specified directory.
}
\description{
The function compiles Rmarkdown Rnw file to PDF document.
}
\note{
None.
}
\examples{
\dontrun{

}
}
\references{
None
}
\author{
Xia Shen
}
