% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcolor.R
\name{ggcolor}
\alias{ggcolor}
\title{Generate colors like ggplot style}
\usage{
ggcolor(n)
}
\arguments{
\item{n}{An integer for the number of colors.}
}
\description{
This function generates a number of ggplot style colors
}
\examples{
\dontrun{

ggcolor(8)

}
}
\author{
Xia Shen
}
