% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepinfile.R
\name{grepinfile}
\alias{grepinfile}
\title{Find text in all the files in a directory}
\usage{
grepinfile(x, loc = "./")
}
\arguments{
\item{x}{A specific pattern to be found.}

\item{loc}{A string giving the directory to search. Defaults to the current directory.}
}
\value{
A series of grep results will be printed on the screen.
}
\description{
The function finds a specific pattern of text in a specific directory.
}
\note{
None.
}
\examples{
\dontrun{

grepinfile('a')

}
}
\references{
None
}
\author{
Xia Shen
}
