% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{gwas}
\alias{gwas}
\title{The function conducts GWAS or multiple linear regression}
\usage{
gwas(y, X)
}
\arguments{
\item{y}{A vector of the geneotype of each individual.}

\item{X}{A matrix or data.frame of several phenotypes or one phenotypes of each individual,requring X must be an array of at least two dimensions}
}
\value{
A data.fame of beta, se, p-vaule
}
\description{
The function conducts GWAS or multiple linear regression
}
\note{
you can also use this fuction to conduct linear regression.
}
\examples{

a <- sample(c(0,1,2),10,replace=T)	
b <- matrix(rnorm(50,2,1),10,5)
gwas(a,b)

}
\author{
Ting Li, Xia Shen
}
