% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hp.R
\name{hp}
\alias{hp}
\title{Drawing heatmap}
\usage{
hp(
  x,
  output.path = "~/Desktop/heatmap.pdf",
  mode = "matrix",
  x.lab = NA,
  y.lab = NA,
  pmatrix = NA,
  legend = "-log10(P)",
  low.color = 10,
  high.color = 1,
  mid.color = "white",
  title = NA,
  basepoint = 0,
  pdf.width = NA,
  pdf.height = NA,
  axis.lab = F,
  legend.break = NA,
  sig.level = 0.05,
  sig.lab = "*"
)
}
\arguments{
\item{x}{a matrix of values to be ploted.}

\item{output.path}{a character of the output pdf file path. Default is '~/Desktop/heatmap.pdf'.}

\item{mode}{It indicates the mode of heatmap. It should be 'matrix' or 'melt'. If mode = 'matrix', please input your raw matrix. If you want to input data in melt form after melt() function of reshape2 package.}

\item{x.lab}{a character of x axis label. Default is NA.}

\item{y.lab}{a character of y axis label. Default is NA.}

\item{pmatrix}{a matrix of p-values, not necessary.}

\item{legend}{a character of legend name. Default is '-log10(P)'.}

\item{low.color}{a integer indicates the color you want to use when value is low. Default is 10.}

\item{high.color}{a integer indicates the color you want to use when value is high. Default is 1.}

\item{mid.color}{a character indicates the color you want to use for mid-value. Default is 'white'.}

\item{title}{a character of figure title. Default is NA.}

\item{basepoint}{a integer of your mid-value. Default is 0.}

\item{pdf.width}{a integer of width of the output pdf. Default is NA for automatic setting.}

\item{pdf.height}{a integer of height of the output pdf. Default is NA for automatic setting.}

\item{axis.lab}{a logical value decides whether label of axis to be shown. Default is FALSE.}

\item{legend.break}{a vector of break points for legend. Default is NA.}

\item{sig.level}{significant levels of pmatrix(p-values). Default is 0.05.}

\item{sig.lab}{label of significant results. Default is '*'.}
}
\value{
A pdf file for heatmap produced in \code{output.path}.
}
\description{
This function is an user-friendly and automatic heatmap drawing function.
}
\note{
Please change the parameter \code{output.path} to makesure the output pdf file will be produced in your target directory. The parameters only need to be set is \code{x} the matrix you want to draw and \code{output.path} for a very easy using.
If you want to draw something like enrichment values, and highlight the significant results, you can input the enrichment values matrix as \code{x} and input the p-values matrix as \code{pmatrix}. And also you can set significant levels of p-values as \code{sig.level}.
}
\examples{
\dontrun{

set.seed(666)
a <- matrix(rnorm(30,sd = 3),ncol = 5)
hp(a)
# null device 
#          1 

p <- matrix(runif(30),ncol = 5)
hp(a,output.path = "~/Downloads//heatmap.pdf",pmatrix = p)
# null device 
#          1 
# Warning message:
# Removed 29 rows containing missing values (geom_text). 
}

}
\references{
None
}
\seealso{
None
}
\author{
Zhijian Yang
}
