% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.packages.rforge.R
\name{install.packages.rforge}
\alias{install.packages.rforge}
\title{Install package from R-Forge}
\usage{
install.packages.rforge(x)
}
\arguments{
\item{x}{A string giving the package name.}
}
\value{
None.
}
\description{
The function installs packages from the R-Forge.
}
\note{
None.
}
\examples{
\dontrun{

install.packages.rforge('sysubiostat')

}
}
\references{
None
}
\author{
Xia Shen
}
