% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll.R
\name{ll}
\alias{ll}
\title{The function checks the existed files in your current working directory}
\usage{
ll()
}
\value{
returns a list of files in your current working directory.
}
\description{
The function checks the existed files in your current working directory
}
\note{
Please use ll() to use this function.
}
\examples{

ll()

}
\author{
Ting Li, Xia Shen
}
