% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lo.R
\name{lo}
\alias{lo}
\title{Summarize all objects in the workspace}
\usage{
lo(pos = 1, pattern, order.by = 2, decreasing = TRUE, head = TRUE, n = 99)
}
\arguments{
\item{pos}{See ?ls pos argument. Default = 1 for the used input objects.}

\item{pattern}{See ?ls pattern argument. List objects with specified pattern.}

\item{order.by}{Ordering by which column. 1: Class; 2 (Default): Size; 3: Row number; 4: Column number.}

\item{decreasing}{Whether to list in decreasing order. Default: \code{TRUE}.}

\item{head}{Whether to head the list. Default: \code{TRUE}.}

\item{n}{Number of objects to be considered in \code{head}.}
}
\value{
A summary list will be printed on the screen.
}
\description{
The function summarizes all the objects in the workspace in a human readable format.
}
\note{
Normally no argument need to be specified.
}
\examples{
\dontrun{

lo()

}
}
\references{
None
}
\author{
Xia Shen
}
