% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nejm.colors.R
\name{nejm.colors}
\alias{nejm.colors}
\title{nejm style colors}
\usage{
nejm.colors(n = 8, alpha = 1)
}
\arguments{
\item{n}{An integer for the number of colors,  default 8,  required <=8.}

\item{alpha}{An numeric for alpha, default 1, requied >=0 and <=1.}
}
\description{
The function generates colors character for nejm likely
}
\examples{
\dontrun{

nejm.colors(7,0.2)

}
}
\author{
Xia Shen
}
