% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.data.frame.R
\name{write.data.frame}
\alias{write.data.frame}
\title{Write Data Frame}
\usage{
write.data.frame(x, fn)
}
\arguments{
\item{x}{the object to be written, preferably a matrix or data frame. If not, it is attempted to coerce x to a data frame.}

\item{fn}{a character string naming a file}
}
\value{
prints x to a file named by fn
}
\description{
The function writes data frame with its colnames to a file, using '\t' as filed separator.
}
\note{
write.data.frame() does not return rownames of x. No columns will be surrounded by double quotes.
}
\examples{

}
\references{

}
\seealso{
write.matrix()
}
\author{
Xia Shen
}
