% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.matrix.R
\name{write.matrix}
\alias{write.matrix}
\title{Write Matrix}
\usage{
write.matrix(x, fn)
}
\arguments{
\item{x}{the object to be written, preferably a matrix or data frame. If not, it is attempted to coerce x to a data frame.}

\item{fn}{a character string naming a file}
}
\value{
prints x to a file named by fn
}
\description{
The function writes data frame without colnames to a file, using '\t' as filed separator.
}
\note{
write.matrix() does not return rownames or colnames of x. No columns will be surrounded by double quotes.
}
\examples{

}
\references{

}
\seealso{
write.data.frame()
}
\author{
Xia Shen
}
