% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore.R
\name{zscore}
\alias{zscore}
\title{The function estimates z-score of specific Random Variables, which measure the standard deviations from the mean}
\usage{
zscore(x)
}
\arguments{
\item{x}{specific Random Variables}
}
\value{
standard z-score
}
\description{
The function estimates z-score of specific Random Variables, which measure the standard deviations from the mean
}
\note{
Please refer to the documentation of the \code{optim} function for method details.
}
\examples{
\dontrun{

num <- c(1051, 179, 1028, 154, 1040, 159, 981, 208) 

zscore (x=num) 

}
}
\references{
Yang Z, Xu W, Zhai R, Li T, Ning Z, Pawitan Y, Shen X (2020). Triangulation of analysis strategies links complex traits to specific tissues and cell types. 
\emph{Submitted}.
}
\seealso{
\code{combined.fdr}
}
\author{
Xia Shen
}
